Fix potential linker error and warning
